requirements.txt
python-dotenv==1.1.1
aiogram==3.4.1
aiohttp==3.9.1
requests==2.31.0
aiofiles==23.2.1